requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==7.1.2
Flask==1.1.4
flask-restx==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.3
pyrsistent==0.18.0
pytz==2021.1
six==1.16.0
Werkzeug==1.0.1
